// apps/creator-hub/modules/permissions/utils/enums.ts
export enum EntityTypes {
  ORGANIZATION = 'Organization',
  UNIVERSE = 'Universe',
}
export enum CreatorTypes {
  MEMBER_ROLE = 'Member',
  ROLE = 'Role',
  USER = 'User',
  LEGACY_ROLE = 'LegacyRole',
  GROUP = 'Group',
}

export enum CreatorFilterChipTypes {
  ALL = 'All',
  USER = 'User',
  ROLE = 'Role',
}